% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve-write-vetiver.R
\name{valve_write_vetiver}
\alias{valve_write_vetiver}
\title{Write a Dockerfile for vetiver using valve}
\usage{
valve_write_vetiver(
  vetiver_model,
  plumber_file = "plumber.R",
  path = ".",
  port = 8000,
  vetiver_args = list()
)
}
\arguments{
\item{vetiver_model}{A deployable \code{\link[vetiver:vetiver_model]{vetiver_model()}} object}

\item{plumber_file}{A path for your Plumber file, created via
\code{\link[vetiver:vetiver_write_plumber]{vetiver_write_plumber()}}. Defaults to \code{plumber.R} in the working directory.}

\item{path}{A path to write the Dockerfile and \code{lockfile}, capturing the
model's package dependencies. Defaults to the working directory.}

\item{port}{The server port for listening: a number such as 8080 or an
expression like \code{'as.numeric(Sys.getenv("PORT"))'} when the port is injected
as an environment variable.}

\item{vetiver_args}{additional arguments passed to \code{vetiver::vetiver_write_docker()} as
key-value pairs in a list object.}
}
\value{
The content of the Dockerfile, invisibly.
}
\description{
This function is a modification of \code{vetiver::vetiver_write_docker()} which utilizes
valve to run model prediction endpoints in parallel.
}
